6DOF_v4_pure/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
pathlib
arabic_reshaper>=3.0.0
python-bidi>=0.4.2
PyYAML>=5.4.0
gymnasium>=0.29.0
stable-baselines3>=2.3.0
sb3-contrib>=2.3.0
torch>=2.0.0
websockets>=12.0
tensorboard>=2.14.0
